Moved embed modal functionality into separate folder, created main modal js file
