--- js/modals/modal-shared.js
@@ -1,2 +1,0 @@
-
-
